Modularized repo, changed to lib, added testing
